requirements.txt
certifi==2017.11.5
chardet==3.0.4
coverage==4.4.2
dbus-python==1.2.4
idna==2.6
nose==1.3.7
nose-testconfig==0.10
PyYAML==3.12
requests==2.18.4
ruamel.yaml==0.15.35
urllib3==1.22
vext==0.7.0
vext.gi==0.6.8
